nodetypes: Increased integer size
backward_length was overflowing in some nodes

+++ nodetypes.h
@@ -1,10 +1,10 @@
 #include <inttypes.h>
 struct wikiNode {
-    int8_t dist_a;
-    int8_t dist_b;
-    int16_t forward_length;
-    int16_t backward_length;
-    int32_t references[];
+    uint8_t dist_a;
+    uint8_t dist_b;
+    uint16_t forward_length;
+    uint32_t backward_length;
+    uint32_t references[];
 };
 // this struct should be made @ runtime by combining title and node data
 // title variable type might change to int32_t
